requirements.txt
openai>=1.0.0
requests>=2.31.0
sendgrid>=6.10.0
python-dotenv>=1.0.0
